fix 3 task, swithc int to double

+++ src/main/java/day3/Task3.java
@@ -7,13 +7,13 @@ public class Task3 {
         Scanner sc = new Scanner(System.in);
 
         for (int i = 0; i < 5; i++) {
-            int divident = sc.nextInt();
-            int divisor = sc.nextInt();
+            double divident = sc.nextDouble();
+            double divisor = sc.nextDouble();
             if (divisor == 0) {
                 System.out.println("Деление на 0");
                 continue;
             }
-            int result = divident / divisor;
+            double result = divident / divisor;
             System.out.println(result);
         }
     }
